pyproject.toml
[project]
name = "mcp-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "fastmcp>=2.11.2",
    "ffmpeg-python>=0.2.0",
    "markdownify>=1.1.0",
    "openai-whisper>=20250625",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
    "uvicorn>=0.35.0",
    "yt-dlp>=2025.7.21",
]
